fetch spec from url provided in query parameter
